change csproj about ThreadPoolServer compile remove
